requirements.txt
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
commonmark==0.9.1
cryptography==36.0.0
debugpy==1.5.1
decorator==5.1.0
discord-webhook==0.14.0
entrypoints==0.3
h11==0.12.0
idna==3.3
ipykernel==6.6.0
ipython==7.30.1
jedi==0.18.1
jupyter-client==7.1.0
jupyter-core==4.9.1
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
outcome==1.1.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.23
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.10.0
pyOpenSSL==21.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pyzmq==22.3.0
requests==2.26.0
rich==10.15.2
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tornado==6.1
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.7
wcwidth==0.2.5
wsproto==1.0.0
